creaded GlobalChat
